{{cookiecutter.app_name}}/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest>=2.6.3
factory-boy==2.5.1

# Management script
Flask-Script
